Back to normal
Changed the whole method back to the commit that actually worked,
BigDecimal...
It may be slower, but I didn't find any **proper** solution to fix the
floating point issues. This will work and that's the whole point of it.
BigDecimal is bad to use if you're using it to constantly calculate
values, but now you are just using it when someone actually uses this
expression.