Fix slow variable saving
Fixes slow variable saving.

As per the desearialize method of the OfflinePlayer serializer, it only requires the uuid field when the server supports UUIDs. Hence making the name field useless when UUIDs are supported on the server. As per why this makes variable saving slow, the UUID of every OfflinePlayer is known instantly, but the name data requires getting it from the world data (offline-mode) or Mojang servers (online-mode), and that makes variable saving slow.

This may seem just a two line change nothing much but its affect is huge if you have many list variables deeply-nested with player expressions on them, storing serialized OfflinePlayers. My 5 year old server had 730.000+ variables with many of them being like that (don't ask me why i had that many variables, i was writing scripts inexperienced) and it has been shutdowning in 8-10 minutes. After this change, it went to 10 seconds.

As for how I found this bottleneck, at first I added a variable save tracker like the one that says "Loaded .. variables so far" when loading variables that says "Saved ... variables so far". This confirmed that the freeze when disabling Skript is from variables. And after that I coded a custom implementation of Paper's spike/freeze detector (which also copies Spigot's built-in timeout detector) that also works on shutdown. Paper's and Spigot's timeout/spike/freeze detectors are disabled when plugins are loading or disabling since they all do update checks or saves and the spikes/freezes are expected.

This does not only occur on shutdown since Skript also periodically saves variables to the database on the runtime, too. The runtime saving performance was improved a lot with the commits a30acb9 ,  2b126c7 and bf4abf6 by bensku, however, this will still increase runtime save performance too, i think. The bottleneck may be also cause of many TPS drops or such, but it was only seen on shutdown since it force saves variables and force waits it to end as a primary task, as the runtime save will run when the server is running, i was probably just not noticing it.

Note that, even though I tested this on my server with 730.000+ variables; i didn't test it on production with users. My server was closed long ago, it was just a hobby for me to figure out why it was closing like that when it was still alive. I was thinking this was normal because of high variable amount that I had back in the days. Users should backup their variables before testing this commit/version.

Signed-off-by: Mustafa ÖNCEL <support@lifemcserver.com>